requirements.txt
alembic==1.13.1
asttokens==2.4.1
audioread==3.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
greenlet==3.0.3
idna==3.4
intel-openmp==2021.4.0
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
learn==1.0.0
librosa==0.10.2.post1
llvmlite==0.42.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
opencv-python==4.10.0.84
optuna==3.6.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pooch==1.8.1
prompt_toolkit==3.0.47
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sk==0.0.1
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.30
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.13
wcwidth==0.2.13